tree: decompose `cass_types` module
The 3 remaining items there are just public re-exports. They are now
moved to other (reasonable) modules, and `cass_types` module is deleted.
