--- scylla-rust-wrapper/src/cass_types.rs
@@ -1,3 +1,0 @@
-pub use crate::cass_batch_types::CassBatchType;
-pub use crate::cass_consistency_types::CassConsistency;
-pub use crate::cass_data_types::CassValueType;
