+++ scylla-rust-wrapper/src/cql_types/data_type.rs
@@ -1,7 +1,7 @@
 use crate::argconv::*;
 use crate::batch::CassBatchType;
 use crate::cass_error::CassError;
-use crate::cass_types::CassValueType;
+use crate::cql_types::CassValueType;
 use crate::types::*;
 use scylla::cluster::metadata::{CollectionType, NativeType};
 use scylla::frame::response::result::ColumnType;
