+++ scylla-rust-wrapper/src/exec_profile.rs
@@ -20,11 +20,11 @@ use crate::argconv::{
 };
 use crate::batch::CassBatch;
 use crate::cass_error::CassError;
-use crate::cass_types::CassConsistency;
 use crate::cluster::{
     set_load_balance_dc_aware_n, set_load_balance_rack_aware_n, update_comma_delimited_list,
 };
 use crate::config_value::{MaybeUnsetConfig, RequestTimeout};
+use crate::cql_types::CassConsistency;
 use crate::load_balancing::{LoadBalancingConfig, LoadBalancingKind};
 use crate::retry_policy::CassRetryPolicy;
 use crate::session::CassConnectedSession;
@@ -869,14 +869,14 @@ mod tests {
     use super::*;
 
     use crate::argconv::CassPtr;
+    use crate::cql_types::CassConsistency;
     use crate::retry_policy::{
         cass_retry_policy_downgrading_consistency_new, cass_retry_policy_free,
     };
     use crate::testing::{assert_cass_error_eq, setup_tracing};
     use crate::{
         argconv::{make_c_str, str_to_c_str_n},
-        batch::{cass_batch_add_statement, cass_batch_free, cass_batch_new},
-        cass_types::CassBatchType,
+        batch::{CassBatchType, cass_batch_add_statement, cass_batch_free, cass_batch_new},
         statement::{cass_statement_free, cass_statement_new},
     };
